backend/requirements.txt
# FastAPI and extensions
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy>=1.4.46,<2.0.0
psycopg2-binary==2.9.9

# Database and Cloud SQL
cloud-sql-python-connector==1.2.0
pg8000==1.29.8

# Google Cloud
google-cloud-translate>=3.0.0
google-generativeai==0.8.4

# JWT for authentication
PyJWT==2.8.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0

#ML configs
scikit-learn
pandas
numpy

# Testing
pytest==7.4.0
pytest-cov==4.1.0
